@@ -1,10 +1,10 @@
 module github.com/quasarlabs/quasarnode
 
-go 1.18
+go 1.20
 
 require (
-	github.com/CosmWasm/wasmd v0.31.0
-	github.com/cosmos/cosmos-sdk v0.45.14
+	github.com/CosmWasm/wasmd v0.31.1-0.20230322155833-d48bb69b391e
+	github.com/cosmos/cosmos-sdk v0.45.15
 	github.com/cosmos/ibc-go/v4 v4.3.0
 	github.com/gogo/protobuf v1.3.3
 	github.com/golang/mock v1.6.0
