@@ -84,7 +84,7 @@ func NewRootCmd() (*cobra.Command, appparams.EncodingConfig) {
 
 // initTendermintConfig helps to override default Tendermint Config values.
 // return tmcfg.DefaultConfig if no custom configuration is required for the application.
-//func initTendermintConfig() *tmcfg.Config {
+// func initTendermintConfig() *tmcfg.Config {
 //	cfg := tmcfg.DefaultConfig()
 //	return cfg
 //}
@@ -301,7 +301,7 @@ func (a appCreator) appExport(
 
 // initAppConfig helps to override default appConfig template and configs.
 // return "", nil if no custom configuration is required for the application.
-func initAppConfig() (string, interface{}) {
+func initAppConfig() (string, any) {
 	// The following code snippet is just for reference.
 
 	// WASMConfig defines configuration for the wasm module.
