@@ -98,9 +98,9 @@ func (k Keeper) findOsmosisEpochByIdentifier(ctx sdk.Context, identifier string)
 }
 
 // findGaugeWeight iterates over distrInfo.Records and returns the weight of record is it finds and record with given gaugeId.
-func findGaugeWeight(ctx sdk.Context, gaugeId uint64, distrInfo poolincentivestypes.DistrInfo) (sdk.Int, bool) {
+func findGaugeWeight(ctx sdk.Context, gaugeID uint64, distrInfo poolincentivestypes.DistrInfo) (sdk.Int, bool) {
 	for _, record := range distrInfo.Records {
-		if record.GaugeId == gaugeId {
+		if record.GaugeId == gaugeID {
 			return record.Weight, true
 		}
 	}
