@@ -36,12 +36,12 @@ func DefaultParams() Params {
 }
 
 // ParamSetPairs get the params.ParamSet
-func (p *Params) ParamSetPairs() paramtypes.ParamSetPairs {
+func (*Params) ParamSetPairs() paramtypes.ParamSetPairs {
 	return paramtypes.ParamSetPairs{}
 }
 
 // Validate validates the set of params
-func (p Params) Validate() error {
+func (Params) Validate() error {
 	return nil
 }
 
