@@ -5,6 +5,4 @@ import (
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 )
 
-var (
-	ErrInvalidMetadataFormat = sdkerrors.New(ModuleName, 2, "invalid metadata format")
-)
+var ErrInvalidMetadataFormat = sdkerrors.New(ModuleName, 2, "invalid metadata format")
